refactored slice to a single userSlice.js file. removed seperated tokenSlices.js due to race condition issue.
