+++ src/app/api.js
@@ -1,18 +1,21 @@
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
+import { getToken } from "./tokenService";
+
 const API_URL = "https://fsa-capstone-backend-9b3e597200f5.herokuapp.com";
+
 const api = createApi({
   reducerPath: "api",
   baseQuery: fetchBaseQuery({
     baseUrl: API_URL,
     prepareHeaders: (headers) => {
-      const token = localStorage.getItem("token");
+      const token = getToken();
       if (token) {
         headers.set("Authorization", `Bearer ${token}`);
       }
       return headers;
     },
   }),
-  tagTypes: ["User", "Auth"],
+  tagTypes: ["User"],
   endpoints: () => ({}),
 });
 export default api;
