+++ src/app/store.js
@@ -1,11 +1,12 @@
 import { configureStore } from "@reduxjs/toolkit";
+import { registerReducer, loginReducer } from "./userSlice";
 import api from "./api";
-import reducer from "../components/userSlice";
 
-export const store = configureStore({
+const store = configureStore({
   reducer: {
     [api.reducerPath]: api.reducer,
-    reducer,
+    register: registerReducer,
+    login: loginReducer,
   },
   middleware: (getDefaultMiddleWare) =>
     getDefaultMiddleWare().concat(api.middleware),
